updare random engine
removed engine_rng

* moved engine decalaration into backend_sycl
