+++ dpnp/backend/mkl_wrap_rng.cpp
@@ -32,7 +32,6 @@
 #include <backend_iface.hpp>
 #include "backend_utils.hpp"
 #include "queue_sycl.hpp"
-#include "rng_engine.hpp"
 
 namespace mkl_rng = oneapi::mkl::rng;
 
@@ -52,16 +51,8 @@ void mkl_rng_gaussian(void* result, size_t size)
 
     mkl_rng::gaussian<_DataType> distribution(mean, stddev);
     // perform generation
-    try
-    {
-        mkl_rng::generate(distribution, DPNP_RNG_ENGINE, size, result1);
-        DPNP_QUEUE.wait_and_throw();
-    }
-    catch (cl::sycl::exception const& e)
-    {
-        std::cerr << "Caught synchronous SYCL exception during mkl_rng_uniform_mt19937():\n"
-                  << e.what() << "\nOpenCL status: " << e.get_cl_code() << std::endl;
-    }
+    auto event_out = mkl_rng::generate(distribution, DPNP_RNG_ENGINE, size, result1);
+    DPNP_QUEUE.wait();
 }
 
 template <typename _DataType>
@@ -79,17 +70,10 @@ void mkl_rng_uniform(void* result, long low, long high, size_t size)
     const _DataType b = (_DataType(high));
 
     mkl_rng::uniform<_DataType> distribution(a, b);
-    try
-    {
-        // perform generation
-        auto event_out = mkl_rng::generate(distribution, DPNP_RNG_ENGINE, size, result1);
-        DPNP_QUEUE.wait_and_throw();
-    }
-    catch (cl::sycl::exception const& e)
-    {
-        std::cerr << "Caught synchronous SYCL exception during mkl_rng_uniform_mt19937():\n"
-                  << e.what() << "\nOpenCL status: " << e.get_cl_code() << std::endl;
-    }
+    // perform generation
+    auto event_out = mkl_rng::generate(distribution, DPNP_RNG_ENGINE, size, result1);
+    DPNP_QUEUE.wait();
+
 }
 
 template void mkl_rng_gaussian<double>(void* result, size_t size);
